requirements.txt
python-telegram-bot==21.0.1
google-generativeai==0.8.0
python-dotenv==1.0.1
Pillow==10.4.0
schedule==1.2.2
pydub==0.25.1
SpeechRecognition==3.10.4
pytz==2024.1
